Rename becomeEditorModal -> requestEditorConfirmModal
