+++ static/src/editors/RequestEditorButton.vue
@@ -26,7 +26,7 @@
             type="submit"
             class="btn btn-gray"
             title="Obtenir les droits de rédaction..."
-            @click="switchToEditorPage"
+            @click="takeEditorRights"
             >
             <i class="fa fa-exchange-alt mr-1"></i>
             <span>Obtenir les droits de rédaction...</span>
@@ -38,15 +38,15 @@
       </error-bar>
     </div>
     <request-editor-modal id="requestEditorModal" :questionnaire="questionnaire"></request-editor-modal>
-    <become-editor-modal id="becomeEditorModal"
-                         :questionnaire-id="questionnaire.id">
-    </become-editor-modal>
+    <request-editor-confirm-modal id="requestEditorConfirmModal"
+                                  :questionnaire-id="questionnaire.id">
+    </request-editor-confirm-modal>
   </div>
 </template>
 
 <script>
 import backendUrls from '../utils/backend.js'
-import BecomeEditorModal from '../editors/BecomeEditorModal'
+import RequestEditorConfirmModal from '../editors/RequestEditorConfirmModal'
 import ErrorBar from '../utils/ErrorBar'
 import { mapFields } from 'vuex-map-fields'
 import RequestEditorModal from '../editors/RequestEditorModal'
@@ -60,8 +60,8 @@ export default Vue.extend({
     }
   },
   components: {
-    BecomeEditorModal,
     ErrorBar,
+    RequestEditorConfirmModal,
     RequestEditorModal,
   },
   computed: {
@@ -74,7 +74,7 @@ export default Vue.extend({
         editor: editorUser,
       })
     },
-    switchToEditorPage: function() {
+    takeEditorRights: function() {
       this.errorMessage = ''
       this.callSwapEditorApi(this.sessionUser.id, this.questionnaire.id)
         .then((response) => {
