+++ static/src/editors/RequestEditorConfirmModal.vue
